+++ packages/language-support/src/tests/signatureHelp.test.ts
@@ -1,10 +1,10 @@
-import {
-  ParameterInformation,
-  SignatureHelp,
-  SignatureInformation,
-} from 'vscode-languageserver-types';
+import { SignatureHelp } from 'vscode-languageserver-types';
 import { DbSchema } from '../dbSchema';
-import { emptyResult, signatureHelp } from '../signatureHelp';
+import {
+  emptyResult,
+  signatureHelp,
+  toSignatureInformation,
+} from '../signatureHelp';
 import { testData } from './testData';
 
 export function testSignatureHelp(
@@ -23,13 +23,7 @@ export function testSignatureHelp(
 describe('Procedures signature help', () => {
   const dbSchema = testData.mockSchema;
   const procedureName = 'apoc.do.when';
-  const signature = SignatureInformation.create(
-    procedureName,
-    dbSchema.procedureSignatures[procedureName].documentation,
-    ...dbSchema.procedureSignatures[procedureName].parameters.map((param) =>
-      ParameterInformation.create(param.label, param.documentation),
-    ),
-  );
+  const signature = toSignatureInformation(dbSchema.procedures[procedureName]);
 
   function expectedArgIndex(i: number): SignatureHelp {
     return {
@@ -219,18 +213,17 @@ describe('Procedures signature help', () => {
       63,
     );
   });
+
+  test('Does not crash on missing schema', () => {
+    testSignatureHelp(`CALL apoc.do.when`, undefined, emptyResult);
+    testSignatureHelp(`CALL apoc.do.when`, {}, emptyResult);
+  });
 });
 
 describe('Functions signature help', () => {
   const dbSchema = testData.mockSchema;
   const functionName = 'apoc.coll.combinations';
-  const signature = SignatureInformation.create(
-    functionName,
-    dbSchema.functionSignatures[functionName].documentation,
-    ...dbSchema.functionSignatures[functionName].parameters.map((param) =>
-      ParameterInformation.create(param.label, param.documentation),
-    ),
-  );
+  const signature = toSignatureInformation(dbSchema.functions[functionName]);
 
   function expectedArgIndex(i: number): SignatureHelp {
     return {
@@ -390,5 +383,10 @@ describe('Functions signature help', () => {
       expectedArgIndex(1),
       89,
     );
+  });
+
+  test('Does not crash on missing schema', () => {
+    testSignatureHelp(`CALL apoc.do.when`, undefined, emptyResult);
+    testSignatureHelp(`CALL apoc.do.when`, {}, emptyResult);
   });
 });